+++ pkg/kernel/src/interrupt/syscall/mod.rs
@@ -8,7 +8,7 @@ use super::consts;
 use service::*;
 
 pub unsafe fn reg_idt(idt: &mut InterruptDescriptorTable) {
-    idt[consts::Interrupts::Syscall as usize]
+    idt[consts::Interrupts::Syscall as u8]
         .set_handler_fn(syscall_handler)
         .set_stack_index(gdt::SYSCALL_IST_INDEX)
         .set_privilege_level(x86_64::PrivilegeLevel::Ring3);
